# Improve indentation performance