@ohler55 2022-03-06
Slight loss on the most common indentation of 2 spaces but significant improvement on more than 2 so looks good to me.